feat: update source index type from u8 to u16 in command and client message enums

+++ src/engine/commands.rs
@@ -3,6 +3,9 @@ use atrium_core::speaker::RenderMode;
 
 /// Commands sent from the control thread to the audio thread via rtrb ring buffer.
 /// All variants must be small and Copy — no heap allocations.
+///
+/// Source indices are `u16` to accommodate both real sources and virtual sources
+/// (e.g. reflection images from ray-traced audio), which can multiply quickly.
 #[derive(Clone, Copy, Debug)]
 pub enum Command {
     /// Update the listener's position and orientation.
@@ -12,10 +15,10 @@ pub enum Command {
     SetMasterGain { gain: f32 },
 
     /// Mute or unmute a source by index.
-    SetSourceMuted { index: u8, muted: bool },
+    SetSourceMuted { index: u16, muted: bool },
 
     /// Reposition a source by index.
-    SetSourcePosition { index: u8, position: Vec3 },
+    SetSourcePosition { index: u16, position: Vec3 },
 
     /// Switch rendering mode (SpeakerAsMic or Vbap).
     SetRenderMode { mode: RenderMode },
@@ -24,16 +27,16 @@ pub enum Command {
     SetSpeakerPosition { channel: u8, position: Vec3 },
 
     /// Set MDAP spread for a source (0.0 = point, 1.0 = full hemisphere).
-    SetSourceSpread { index: u8, spread: f32 },
+    SetSourceSpread { index: u16, spread: f32 },
 
     /// Set orbit speed for a source (0 = paused).
-    SetSourceOrbitSpeed { index: u8, speed: f32 },
+    SetSourceOrbitSpeed { index: u16, speed: f32 },
 
     /// Set orbit radius for a source.
-    SetSourceOrbitRadius { index: u8, radius: f32 },
+    SetSourceOrbitRadius { index: u16, radius: f32 },
 
     /// Set orbit angle for a source.
-    SetSourceOrbitAngle { index: u8, angle: f32 },
+    SetSourceOrbitAngle { index: u16, angle: f32 },
 
     /// Set atmospheric conditions for ISO 9613-1 air absorption.
     SetAtmosphere { temperature_c: f32, humidity_pct: f32 },
